+++ src/main/java/com/yeta/pps/vo/SellApplyOrderVo.java
@@ -7,7 +7,6 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -132,13 +131,13 @@ public class SellApplyOrderVo {
      * 总商品金额
      */
     @NotNull(message = CommonResponse.MESSAGE3)
-    private BigDecimal totalMoney;
+    private Double totalMoney;
 
     /**
      * 直接优惠金额
      */
     @NotNull(message = CommonResponse.MESSAGE3)
-    private BigDecimal discountMoney;
+    private Double discountMoney;
 
     /**
      * 优惠券编号
@@ -149,23 +148,23 @@ public class SellApplyOrderVo {
      * 总优惠金额
      */
     @NotNull(message = CommonResponse.MESSAGE3)
-    private BigDecimal totalDiscountMoney;
+    private Double totalDiscountMoney;
 
     /**
      * 本单金额
      */
     @NotNull(message = CommonResponse.MESSAGE3)
-    private BigDecimal orderMoney;
+    private Double orderMoney;
 
     /**
      * 已结算金额
      */
-    private BigDecimal clearedMoney;
+    private Double clearedMoney;
 
     /**
      * 未结算金额
      */
-    private BigDecimal notClearedMoney;
+    private Double notClearedMoney;
 
     /**
      * 经手人编号
@@ -410,19 +409,19 @@ public class SellApplyOrderVo {
         this.outNotSentQuantity = outNotSentQuantity;
     }
 
-    public BigDecimal getTotalMoney() {
+    public Double getTotalMoney() {
         return totalMoney;
     }
 
-    public void setTotalMoney(BigDecimal totalMoney) {
+    public void setTotalMoney(Double totalMoney) {
         this.totalMoney = totalMoney;
     }
 
-    public BigDecimal getDiscountMoney() {
+    public Double getDiscountMoney() {
         return discountMoney;
     }
 
-    public void setDiscountMoney(BigDecimal discountMoney) {
+    public void setDiscountMoney(Double discountMoney) {
         this.discountMoney = discountMoney;
     }
 
@@ -434,35 +433,35 @@ public class SellApplyOrderVo {
         this.discountCouponId = discountCouponId;
     }
 
-    public BigDecimal getTotalDiscountMoney() {
+    public Double getTotalDiscountMoney() {
         return totalDiscountMoney;
     }
 
-    public void setTotalDiscountMoney(BigDecimal totalDiscountMoney) {
+    public void setTotalDiscountMoney(Double totalDiscountMoney) {
         this.totalDiscountMoney = totalDiscountMoney;
     }
 
-    public BigDecimal getOrderMoney() {
+    public Double getOrderMoney() {
         return orderMoney;
     }
 
-    public void setOrderMoney(BigDecimal orderMoney) {
+    public void setOrderMoney(Double orderMoney) {
         this.orderMoney = orderMoney;
     }
 
-    public BigDecimal getClearedMoney() {
+    public Double getClearedMoney() {
         return clearedMoney;
     }
 
-    public void setClearedMoney(BigDecimal clearedMoney) {
+    public void setClearedMoney(Double clearedMoney) {
         this.clearedMoney = clearedMoney;
     }
 
-    public BigDecimal getNotClearedMoney() {
+    public Double getNotClearedMoney() {
         return notClearedMoney;
     }
 
-    public void setNotClearedMoney(BigDecimal notClearedMoney) {
+    public void setNotClearedMoney(Double notClearedMoney) {
         this.notClearedMoney = notClearedMoney;
     }
